se corrigieron errores de logica

+++ src/main/java/com/apirest/apirest/Servicio/Impl/CategoriaService.java
@@ -3,6 +3,7 @@ package com.apirest.apirest.Servicio.Impl;
 import java.util.List;
 import java.util.Optional;
 
+import com.apirest.apirest.Exception.noHayContenidoException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,45 +25,50 @@ public class CategoriaService implements ICategoria{
     @Autowired
     private CategoriaRepository categoriaRepo;
     
-    
 
     @Override
     @Transactional
     public List<categoriaDTO> findAll() {
-        List<categoriaDTO>lista=mapper.toCategoriasDto(categoriaRepo.findAll());
-        return lista;
+        List<categoria>lista=categoriaRepo.findAll();
+        if (lista.isEmpty()){
+            throw new noHayContenidoException();
+        }
+        return mapper.toCategoriasDto(lista);
     }
 
     @Override
-    public Optional<categoriaDTO> findById(Long id) {
+    public categoriaDTO findById(Long id) {
         categoria category=categoriaRepo.findById(id).orElseThrow(()->new CategoriaException());
-        categoriaDTO categoria=mapper.toCategoriaDto(category);
-        return Optional.of(categoria);
+        return mapper.toCategoriaDto(category);
     }
 
     @Override
     public void save(categoriaDTO categoriaDTO) {
-        categoriaRepo.save(mapper.toCategoria(categoriaDTO));
+       categoria categoria=new categoria();
+       categoria.setNombreCategoria(categoriaDTO.getNombre());
+       categoriaRepo.save(categoria);
     }
 
     @Override
-    public Optional<categoriaDTO> findByName(String name) {
+    public categoriaDTO findByName(String name) {
        categoria category=categoriaRepo.findByNombreCategoriaIgnoreCase(name).orElseThrow(()->new CategoriaException());
-       categoriaDTO categoria=mapper.toCategoriaDto(category);
-       return Optional.of(categoria);
-       
-       
+       return mapper.toCategoriaDto(category);
     }
 
     @Override
     public void delete(Long id) {
-      categoriaRepo.deleteById(id);
+        if (categoriaRepo.existsById(id)){
+            categoriaRepo.deleteById(id);
+        }else{
+            throw new CategoriaException();
+        }
+
     }
 
     @Override
-    public categoriaDTO update(Long id, categoriaDTO categoriaDTO) {
+    public categoriaDTO update(Long id, categoriaDTO categoriaDto) {
         categoria category=categoriaRepo.findById(id).orElseThrow(()->new CategoriaException());
-        category.setNombreCategoria(categoriaDTO.getNombre());
+        category.setNombreCategoria(categoriaDto.getNombre());
         categoriaRepo.save(category);
         return mapper.toCategoriaDto(category);
     }
